src/requirements.txt
google-cloud-aiplatform==1.80.0
google-cloud-storage==2.19.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.6.1
joblib==1.4.2
fsspec==2025.2.0
gcsfs==2025.2.0